Create roles. Add testing to new Dao. Manage roles on web page.

+++ Model/Sql/createTables.sql
@@ -130,14 +130,33 @@ IF  EXISTS (SELECT * FROM sys.objects WHERE object_id = OBJECT_ID('[User]')
 AND type in ('U')) DROP TABLE [User]
 GO
 
+/* Drop Table Role if already exists */
+
+IF  EXISTS (SELECT * FROM sys.objects WHERE object_id = OBJECT_ID('[Role]') 
+AND type in ('U')) DROP TABLE [Role]
+GO
+
 /*
  * Create tables.
  */
+
+ /* Role */
+
+CREATE TABLE [Role] (
+	roleId bigint IDENTITY(1,1) NOT NULL,
+	name varchar(10) NOT NULL,
+
+	CONSTRAINT [PK_Role] PRIMARY KEY (roleId ASC)
+)
+
+PRINT N'Table Role created.'
+GO
 
 /* User */
 
 CREATE TABLE [User] (
 	userId bigint IDENTITY(1,1) NOT NULL,
+	roleId bigint NOT NULL,
 	login varchar(30) NOT NULL,
 	password varchar(64) NOT NULL,
 	name varchar(30) NOT NULL,
@@ -147,7 +166,9 @@ CREATE TABLE [User] (
 	language varchar(2) NOT NULL,
 	country varchar(2) NOT NULL,
 
-	CONSTRAINT [PK_User] PRIMARY KEY (userId ASC)
+	CONSTRAINT [PK_User] PRIMARY KEY (userId ASC),
+	CONSTRAINT [FK_UserRoleId] FOREIGN KEY(roleId)
+		REFERENCES [Role] (roleId) ON DELETE CASCADE
 )
 
 PRINT N'Table User created.'
